DB & Model Updates

+++ src/SFA.DAS.Payments.Model.Core/Entities/PaymentModel.cs
@@ -46,6 +46,6 @@ namespace SFA.DAS.Payments.Model.Core.Entities
         public long? ApprenticeshipPriceEpisodeId { get; set; }
         public ApprenticeshipEmployerType ApprenticeshipEmployerType { get; set; }
         public string ReportingAimFundingLineType { get; set; }
-        public int? AimSeqNumber { get; set; }
+        public long? AimSeqNumber { get; set; }
     }
 }
